NPM_CLEAN

+++ src/pa-npm-setup.js
@@ -13,7 +13,7 @@ var path = require('path');
 
 try {
   process.chdir(process.env.PA_HOME + '/pa-web');
-  fs.removeSync('node_modules')
+  if(process.env.NPM_CLEAN === 'true') fs.removeSync('node_modules')
   exec('npm i node-sass@3.4.2');
   exec('npm i');
   exec('gulp development');
